Cleaned up code and added more feedback to pay button
