+++ lib/pages/payment_widgets/pay_button.dart
@@ -7,33 +7,31 @@ import '../../api/network/network_helper.dart';
 import '../../api/network/network_service.dart';
 import '../../api/network/static_values.dart';
 import '../../api/requests/payment_requests.dart';
+import '../error_widget.dart';
 import '../payment_page/payment_page.dart';
 
-class PayButton extends StatelessWidget {
-  const PayButton({
+class PayPreviewButton extends StatelessWidget {
+  /// Button to open payment preview page. It has a hero tag for a nice
+  /// animation when opening the payment preview page.
+
+  const PayPreviewButton({
     super.key,
     required this.licencePlate,
-    this.isPreview = false,
     required this.garage,
   });
 
   final LicencePlate licencePlate;
   final Garage garage;
-  final bool isPreview;
 
   void onPressed(BuildContext context) {
-    if (isPreview) {
-      Navigator.of(context).push(
-        ScaleRoute(
-          page: PaymentPage(
-            licencePlate: licencePlate,
-            garage: garage,
-          ),
+    Navigator.of(context).push(
+      ScaleRoute(
+        page: PaymentPage(
+          licencePlate: licencePlate,
+          garage: garage,
         ),
-      );
-    } else {
-      startPaymentSession(licencePlate: licencePlate.licencePlate);
-    }
+      ),
+    );
   }
 
   @override
@@ -42,7 +40,7 @@ class PayButton extends StatelessWidget {
       tag: 'pay_button_${licencePlate.licencePlate}',
       child: ElevatedButton(
           style: ButtonStyle(
-            shape: MaterialStateProperty.all(StadiumBorder()),
+            shape: MaterialStateProperty.all(const StadiumBorder()),
           ),
           onPressed: () => onPressed(context),
           child: const Text('Pay')),
@@ -50,7 +48,82 @@ class PayButton extends StatelessWidget {
   }
 }
 
+class PayButton extends StatefulWidget {
+  /// Button to launch the Stripe checkout url. It has a hero tag for a nice
+  /// animation when opening the payment preview page.
+
+  const PayButton({
+    super.key,
+    required this.licencePlate,
+    required this.garage,
+  });
+
+  final LicencePlate licencePlate;
+  final Garage garage;
+
+  @override
+  State<PayButton> createState() => _PayButtonState();
+}
+
+class _PayButtonState extends State<PayButton> {
+  Future<void> checkoutFuture = Future(() {});
+
+  void onPressed(BuildContext context) {
+    setState(() {
+      checkoutFuture =
+          startPaymentSession(licencePlate: widget.licencePlate.licencePlate);
+
+      // For testing error state
+      //checkoutFuture = Future.error(Exception('test'));
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Hero(
+      tag: 'pay_button_${widget.licencePlate.licencePlate}',
+      child: FutureBuilder<void>(
+        future: checkoutFuture,
+        builder: (context, snapshot) {
+          if (snapshot.connectionState != ConnectionState.done) {
+            return const Center(
+              child: CircularProgressIndicator(),
+            );
+          } else if (snapshot.hasError) {
+            return Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: [
+                Text(
+                  'We failed to create a checkout page: ${snapshot.error}',
+                  style: const TextStyle(color: Colors.red),
+                ),
+                _buildButton(context, 'Retry')
+              ],
+            );
+          } else {
+            return _buildButton(context, 'Pay');
+          }
+        },
+      ),
+    );
+  }
+
+  Widget _buildButton(BuildContext context, String text) {
+    return FractionallySizedBox(
+      widthFactor: 1,
+      child: ElevatedButton(
+          style: ButtonStyle(
+            shape: MaterialStateProperty.all(const StadiumBorder()),
+          ),
+          onPressed: () => onPressed(context),
+          child: Text(text)),
+    );
+  }
+}
+
 class ScaleRoute extends PageRouteBuilder {
+  /// Route for showing the payment preview page.
   final Widget page;
   ScaleRoute({required this.page})
       : super(
